feat(desktop): use server side ga
