+++ .github/workflows/build-desktop.yml
@@ -32,7 +32,6 @@ env:
   VITE_OPENPANEL_CLIENT_ID: ${{ vars.VITE_OPENPANEL_CLIENT_ID }}
   VITE_OPENPANEL_API_URL: ${{ vars.VITE_OPENPANEL_API_URL }}
   VITE_FIREBASE_CONFIG: ${{ vars.VITE_FIREBASE_CONFIG }}
-  VITE_GA4_KEY: ${{ vars.VITE_GA4_KEY }}
   NODE_OPTIONS: --max-old-space-size=8192
 
 jobs:
